Combine history buttons
